requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
psycopg2-binary
Flask-Login
pytz
fpdf2
Flask-Mail
PyJWT
PyPDF2
docxtpl
ldap3
pytz
pytest
pytest-mock 
unidecode
#pip install -r requirements.txt
